requirements.txt
anyio==3.6.2
arabic-reshaper==2.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.2.5
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
blockchain==1.4.4
cairocffi==1.2.0
CairoSVG==2.5.2
captcha==0.4
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
cssselect2==0.7.0
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.3
django-admin-ckeditor==0.0.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4-datetimepicker==4.2
django-braces==1.13.0
django-ckeditor==5.7.1
django-cors-headers==3.10.1
django-hitcount==1.3.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-nested-admin==4.0.2
django-notifications-hq==1.7.0
django-private-chat==0.2.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
django-tempus-dominus==5.1.2.13
django-upgrade==1.0.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
docutils==0.19
entrypoints==0.4
enum-compat==0.0.3
et-xmlfile==1.0.1
executing==1.2.0
fake-useragent==0.1.11
fastjsonschema==2.16.2
flake8==5.0.4
future==0.18.2
futures==3.1.1
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==4.3.21
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.1.2
jsonfield==3.1.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
Kivy==2.1.0
# kivy-deps.angle
# kivy-deps.glew==0.3.1
# kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.0.2
kiwisolver==1.4.4
lazy-object-proxy==1.4.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
mccabe
mistune==2.0.4
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.0.0
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.0
notifications==0.3.2
ntplib==0.4.0
numpy==1.23.4
openpyxl==2.6.2
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip-review==1.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
pylint
pymongo==3.12.1
PyOpenGL-accelerate==3.1.6
pyparsing==3.0.9
PyPDF2==1.26.0
PyPDF3==1.0.6
pyphen==0.13.0
# pypiwin32==223
pyrsistent==0.19.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
# pywin32==304
# pywinpty==2.0.7
PyYAML==6.0
pyzmq==24.0.1
qrcode==6.1
qtconsole==5.4.0
QtPy==2.2.1
reportlab==3.6.6
requests==2.28.1
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse
stack-data==0.6.0
swapper==1.3.0
sweetify==2.0.0
terminado==0.17.0
tinycss2==1.2.1
tokenize-rt==4.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
# typed-ast==1.4.0
typing_extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
virtualenv==16.7.9
waitress==1.4.4
wcwidth==0.2.5
WeasyPrint==52.4
webencodings==0.5.1
websocket-client==1.4.1
websockets==3.2
widgetsnbextension==4.0.3
# windows-curses==2.3.0
wrapt==1.11.2
xhtml2pdf==0.2.5
zipp==3.10.0
